ai-data-analyst-project/ai-data-analyst-project/requirements.txt
fastapi
uvicorn
pandas
python-multipart
groq
python-dotenv